+++ bluemars/perlin.h
@@ -20,11 +20,10 @@ public:
 protected:
     long seed;
 	std::mt19937 rng; // Random number generator
-    double discreteNoise(long x, long y, int octave);
-    double interpolatedNoise(long x, long y, long waveLength = BASE_WAVE_LENGTH);
+    long discreteNoise(long x, long y, int octave);
+    double interpolatedNoise(long x, long y, int waveLength = BASE_WAVE_LENGTH);
     double cosineInterpolate(long X1, long X2, double Z1, double Z2, long x);
-    
+
 };
 
 #endif
-
